+++ cli/internal_testsuites/typescript/src/testsuite/basic_datastore_and_api_test/basic_datastore_and_api.test.ts
@@ -23,7 +23,7 @@ test("Test basic data store and API", async () => {
 
     if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
 
-    const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+    const { enclaveContext, stopEnclaveFunction, kurtosisContext } = createEnclaveResult.value
 
     try {
         // ------------------------------------- TEST SETUP ----------------------------------------------