+++ cli/internal_testsuites/typescript/src/testsuite/module_test/module.test.ts
@@ -31,7 +31,7 @@ test("Test module", async () => {
 
      if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
  
-     const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+     const { enclaveContext, stopEnclaveFunction, kurtosisContext } = createEnclaveResult.value
  
      try {
         // ------------------------------------- TEST SETUP ----------------------------------------------