+++ cli/internal_testsuites/typescript/src/testsuite/network_partition_test/network_partition.test.ts
@@ -37,7 +37,7 @@ test("Test network partition", async () => {
 
      if(createEnclaveResult.isErr()) { throw createEnclaveResult.error }
  
-     const { enclaveContext, stopEnclaveFunction } = createEnclaveResult.value
+     const { enclaveContext, stopEnclaveFunction, kurtosisContext } = createEnclaveResult.value
  
      try {